+++ src/utils.jl
@@ -55,7 +55,7 @@ end
 
 # Derivative and Gradient
 derivative(a::Measurement{F},
-           tag::Tuple{T, T, Float64}) where {F<:AbstractFloat, T<:AbstractFloat} =
+           tag::Tuple{T, T, UInt64}) where {F<:AbstractFloat, T<:AbstractFloat} =
                get(a.der, tag, zero(F))
 
 """
@@ -104,7 +104,7 @@ Return the components to the uncertainty of the dependent quantity `x` in the
 form of a `Dict`.
 """
 function uncertainty_components(x::Measurement{T}) where {T<:AbstractFloat}
-    out = Dict{Tuple{T, T, Float64}, T}()
+    out = Dict{Tuple{T, T, UInt64}, T}()
     for var in keys(x.der)
         out[var] = abs(var[2] * Measurements.derivative(x, var))
     end
